backend/requirements_new.txt
reportlab>=4.0.0
twilio>=8.0.0
Pillow>=10.0.0
django-filter>=23.0
